Optimized build process
Moved the logic of building the images into docker.  Removed using any
absolute paths on the build system (i.e. $(pwd) which does not work
correctly on Windows)

Added .gitattributes to ensure line endings for files being used in the
Docker build containers are correct.

Please compare with whitespace checks turned off.

In order to support builds within the container bsdtar needs to be used
rather than the regular tar due to
https://github.com/docker/hub-feedback/issues/727

The tar server was also removed in favor of copying from the container
to the "im" folder and imported back using Dockerfile for the final
image.

The step where the files are tarred in install_ihs was removed and put
into docker.  This makes it easier to debug issues with tar